Fix bug where immediate redirects in compat were broken; also save redirect persistence before link and reauth as well
