+++ packages-exp/auth-compat-exp/src/user.ts
@@ -17,6 +17,7 @@
 
 import * as exp from '@firebase/auth-exp/internal';
 import * as compat from '@firebase/auth-types';
+import { _savePersistenceForRedirect } from './persistence';
 import { CompatPopupRedirectResolver } from './popup_redirect';
 import {
   convertConfirmationResult,
@@ -96,7 +97,8 @@ export class User implements compat.User, Wrapper<exp.User> {
       )
     );
   }
-  linkWithRedirect(provider: compat.AuthProvider): Promise<void> {
+  async linkWithRedirect(provider: compat.AuthProvider): Promise<void> {
+    await _savePersistenceForRedirect(exp._castAuth(this.auth));
     return exp.linkWithRedirect(
       this.user,
       provider as exp.AuthProvider,
@@ -144,7 +146,8 @@ export class User implements compat.User, Wrapper<exp.User> {
       )
     );
   }
-  reauthenticateWithRedirect(provider: compat.AuthProvider): Promise<void> {
+  async reauthenticateWithRedirect(provider: compat.AuthProvider): Promise<void> {
+    await _savePersistenceForRedirect(exp._castAuth(this.auth));
     return exp.reauthenticateWithRedirect(
       this.user,
       provider as exp.AuthProvider,
