+++ packages-exp/auth-exp/internal/index.ts
@@ -33,7 +33,7 @@ export {
 } from '../src/model/popup_redirect';
 export { UserCredentialInternal, UserParameters } from '../src/model/user';
 export { registerAuth } from '../src/core/auth/register';
-export { DefaultConfig, AuthImpl } from '../src/core/auth/auth_impl';
+export { DefaultConfig, AuthImpl, _castAuth } from '../src/core/auth/auth_impl';
 
 export { ClientPlatform, _getClientVersion } from '../src/core/util/version';
